[fix] controller interface 삭제 memberId  long->String
