--- back/instagram/src/main/java/com/kbm/instagram/controller/TestControllerImpl.java
@@ -1,16 +1,0 @@
-package com.kbm.instagram.controller;
-
-import com.kbm.instagram.controller.TestController;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/test")
-@CrossOrigin(origins = { "*" }, maxAge = 6000)
-public class TestControllerImpl implements TestController {
-
-    @GetMapping("/getMsg")
-    @ResponseBody
-    public String getMsg() {
-        return "test success!!";
-    }
-}
